mini calendar

+++ pages/api/events/[id].js
@@ -1,5 +1,6 @@
 import dbConnect from '../../../utils/dbConnect'
 import Event from '../../../models/Event'
+const jwt = require('jsonwebtoken')
 
 export default async function handler(req, res) {
   const {
@@ -22,8 +23,15 @@ export default async function handler(req, res) {
       }
       break
 
+      // NEEDS TOKEN
     case 'PUT' /* Edit a model by its ID */:
       try {
+        const token = req.cookies.token
+        const decodedToken = jwt.verify(token, process.env.SECRET)  
+        if (!token || !decodedToken.id) {    
+          return response.status(401).json({ error: 'token missing or invalid' })  
+        }
+
         const event = await Event.findByIdAndUpdate(id, req.body, {
           new: true,
           runValidators: true,
@@ -38,14 +46,23 @@ export default async function handler(req, res) {
       }
       break
 
+      // NEEDS TOKEN
     case 'DELETE' /* Delete a model by its ID */:
-      try {
+       try {
+        const token = req.cookies.token
+        const decodedToken = jwt.verify(token, process.env.SECRET)
+        if (!token || !decodedToken.id) { 
+          console.log('returning error')   
+          return response.status(401).json({ error: 'token missing or invalid' })  
+        }
+        
         const deletedEvent = await Event.deleteOne({ _id: id })
         if (!deletedEvent) {
           return res.status(400).json({ success: false })
         }
         res.status(200).json({ success: true, data: {} })
       } catch (error) {
+        console.log('catching error!')
         res.status(400).json({ success: false })
       }
       break
